+++ config/packages/blopex.py
@@ -48,12 +48,8 @@ class Blopex(package.Package):
       dirs = self.GenerateGuesses('blopex',archdir)
       incdirs = self.GenerateGuesses('blopex',archdir,'include')
 
-    libs = self.packagelibs
-    if not libs:
-      libs = ['-lBLOPEX']
-    includes = self.packageincludes
-    if not includes:
-      includes = ['.']
+    libs = [self.packagelibs] if self.packagelibs else ['-lBLOPEX']
+    includes = [self.packageincludes] if self.packageincludes else ['.']
 
     for (d,i) in zip(dirs,incdirs):
       if d:
@@ -65,13 +61,13 @@ class Blopex(package.Package):
       else:
         l = libs
         f = ['-I' + includes[0]]
-      (result, output) = self.Link([],[],l+f,code,' '.join(f),petsc.language)
+      (result, output) = self.Link([],[],' '.join(l+f),code,' '.join(f),petsc.language)
       if result:
         slepcconf.write('#define SLEPC_HAVE_BLOPEX 1\n')
         slepcvars.write('BLOPEX_LIB = ' + ' '.join(l) + '\n')
         slepcvars.write('BLOPEX_INCLUDE = ' + ' '.join(f) + '\n')
         self.havepackage = True
-        self.packageflags = l+f
+        self.packageflags = ' '.join(l+f)
         self.location = includes[0] if self.packageincludes else i
         return
 
@@ -112,7 +108,7 @@ class Blopex(package.Package):
 
     # Check build
     code = self.SampleCode(petsc)
-    (result, output) = self.Link([],[],[l]+[f],code,f,petsc.language)
+    (result, output) = self.Link([],[],l+' '+f,code,f,petsc.language)
     if not result:
       self.log.Exit('Unable to link with downloaded BLOPEX')
 
@@ -122,5 +118,5 @@ class Blopex(package.Package):
     slepcvars.write('BLOPEX_INCLUDE = ' + f + '\n')
 
     self.havepackage = True
-    self.packageflags = [l] + [f]
+    self.packageflags = l+' '+f